refactor(core): create separate folder for DocumentAPI
