+++ packages/core/src/api/DocumentAPI/index.ts
@@ -1,0 +1,1 @@
+export * from './DocumentAPI.js';
