+++ packages/core/src/api/index.ts
@@ -3,7 +3,7 @@ import { Inject, Service } from 'typedi';
 import { EditorAPI as EditorApiInterface } from '@editorjs/sdk';
 import { BlocksAPI } from './BlocksAPI.js';
 import { SelectionAPI } from './SelectionAPI.js';
-import { DocumentAPI } from './DocumentAPI.js';
+import { DocumentAPI } from './DocumentAPI';
 
 /**
  * Class gathers all Editor's APIs
